Remove use of pkg_resources from ros2cli.
Replace it with the use of the more modern importlib*
libraries.  There are a couple of reasons to do this:

1.  pkg_resources is quite slow to import; on my machine,
just firing up the python interpreter takes ~35ms, while
firing up the python interpreter and importing pkg_resources
takes ~175ms.  Firing up the python interpreter and importing
importlib_metadata takes ~70ms.  Removing 100ms per invocation
of the command-line both makes it speedier for users, and
will speed up our tests (which call out to the command-line
quite a lot).

2.  pkg_resources is somewhat deprecated and being replaced
by importlib.  https://importlib-metadata.readthedocs.io/en/latest/using.html
describes some of it

Note: this change definitively breaks doing releases with
Ubuntu Bionic, as python3-importlib-metadata is not available in Bionic
(though it is available via pip).

Note: By itself, this change is not enough to completely remove our
dependence on pkg_resources.  We'll also have to do something about
the console_scripts that setup.py generates.  That will be
a separate effort.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>